% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_trunc_geom_cdf.R
\name{compute_trunc_geom_cdf}
\alias{compute_trunc_geom_cdf}
\title{Compute the Truncated Geometric Cumulative Distribution Function (CDF)}
\usage{
compute_trunc_geom_cdf(lambda, b)
}
\arguments{
\item{lambda}{A numeric value between 0 and 1 representing the success probability of the geometric distribution.}

\item{b}{An integer specifying the number of support points for truncation. The distribution is truncated at \code{b - 1}.}
}
\value{
A numeric vector of length \code{b} containing the cumulative probabilities of the truncated and normalized geometric distribution.
}
\description{
This function calculates the cumulative distribution function (CDF) for a truncated geometric distribution with a maximum of \code{b} support points.
The geometric distribution is truncated to values from 0 to \code{b - 1}, then normalized.
}
\examples{
compute_trunc_geom_cdf(lambda = 0.2, b = 10)

}
